+++ indexer/services/scripts/src/helpers/block-helpers.ts
@@ -14,21 +14,7 @@ import {
   UpdatePerpetualEventV1,
 } from '@dydxprotocol-indexer/v4-protos';
 
-import { AnnotatedIndexerTendermintEvent } from './types';
-
-export enum DydxIndexerSubtypes {
-  ORDER_FILL = 'order_fill',
-  SUBACCOUNT_UPDATE = 'subaccount_update',
-  TRANSFER = 'transfer',
-  MARKET = 'market',
-  STATEFUL_ORDER = 'stateful_order',
-  FUNDING = 'funding_values',
-  ASSET = 'asset',
-  PERPETUAL_MARKET = 'perpetual_market',
-  LIQUIDITY_TIER = 'liquidity_tier',
-  UPDATE_PERPETUAL = 'update_perpetual',
-  UPDATE_CLOB_PAIR = 'update_clob_pair',
-}
+import { AnnotatedIndexerTendermintEvent, DydxIndexerSubtypes } from './types';
 
 export function annotateIndexerTendermintEvent(
   event: IndexerTendermintEvent,
