+++ indexer/services/scripts/src/helpers/types.ts
@@ -7,3 +7,17 @@ export interface AnnotatedIndexerTendermintEvent extends IndexerTendermintEvent 
 export interface AnnotatedIndexerTendermintBlock extends IndexerTendermintBlock {
   annotatedEvents: AnnotatedIndexerTendermintEvent[];
 }
+
+export enum DydxIndexerSubtypes {
+  ORDER_FILL = 'order_fill',
+  SUBACCOUNT_UPDATE = 'subaccount_update',
+  TRANSFER = 'transfer',
+  MARKET = 'market',
+  STATEFUL_ORDER = 'stateful_order',
+  FUNDING = 'funding_values',
+  ASSET = 'asset',
+  PERPETUAL_MARKET = 'perpetual_market',
+  LIQUIDITY_TIER = 'liquidity_tier',
+  UPDATE_PERPETUAL = 'update_perpetual',
+  UPDATE_CLOB_PAIR = 'update_clob_pair',
+}
